Adds Support for League Median Match Setting #86 (#89)
* Adds Support for League Median Match Setting #86

Sleeper allows leagues to give everyone an extra "matchup"  every week against the league median score.  Score higher than the median, get a win.  Score lower than the median, get a loss.  This commit implements that setting for leagues that have it enabled, so that the standings are accurately reflected in the Standings tab.  As requested/noted in #86

* update league-page version and remove fuzzyjs import

* specify smui version

Co-authored-by: Nicholas Melhado <n.melhado@gmail.com>